refactor: rename MarkFeedback to mark and rename MarkFeedbackHistory to MarkHistory
