+++ api/factories/__init__.py
@@ -3,7 +3,7 @@ from typing import Sequence, Type
 from factory.alchemy import SQLAlchemyModelFactory
 
 from api.factories.answer import AnswerFactory
-from api.factories.mark_feedback import MarkFeedbackFactory, MarkFeedbackHistoryFactory
+from api.factories.mark import MarkFeedbackFactory, MarkFeedbackHistoryFactory
 
 all_factories: Sequence[Type[SQLAlchemyModelFactory]] = [
     AnswerFactory,
