+++ api/router/mark.py
@@ -2,9 +2,9 @@ from fastapi import APIRouter, Depends
 from sqlmodel import Session, select
 
 from api.dependencies import get_assessment_id, get_session
-from api.schemas.mark_feedback import (
-    MarkFeedback,
-    MarkFeedbackRead,
+from api.schemas.mark import (
+    Mark,
+    MarkRead,
 )
 
 mark_feedback_router = APIRouter(prefix="/marks", tags=["marking"])
@@ -12,7 +12,7 @@ mark_feedback_router = APIRouter(prefix="/marks", tags=["marking"])
 
 @mark_feedback_router.get(
     "/{student_username}",
-    response_model=list[MarkFeedbackRead],
+    response_model=list[MarkRead],
     summary="Retrieve user marks and feedback for question",
     description="""
 Retrieve the latest (and historical) marks for answers to a question by user
@@ -25,16 +25,16 @@ def get_marks_feedback_for_question_number(
     assessment_id: str = Depends(get_assessment_id),
 ):
     query = (
-        select(MarkFeedback)
+        select(Mark)
         .where(
-            MarkFeedback.exam_id == assessment_id,
-            MarkFeedback.username == student_username,
+            Mark.exam_id == assessment_id,
+            Mark.username == student_username,
         )
         .order_by(
-            MarkFeedback.question,  # type: ignore
-            MarkFeedback.part,  # type: ignore
-            MarkFeedback.section,  # type: ignore
-            MarkFeedback.task,  # type: ignore
+            Mark.question,  # type: ignore
+            Mark.part,  # type: ignore
+            Mark.section,  # type: ignore
+            Mark.task,  # type: ignore
         )  # type: ignore
     )
     return session.exec(query).all()
